Add files via upload
My aim is to design a generative system which represents the history of the Earth and humanity beginning with the Big Bang at zero point and ending at present date (maybe it ends in the near future). I am going to use a logarihtmic timeline to specify ecospherical and anthropospherical events. So it will comprise almost 14 billion years.
Let's think about a stochastic system that submits observed or learned informations of generation, evolution, developments and modifications through the history. While scaling logaryhtmic timeline to projected real time, I want to integrate these informations as transitions and themes. This will not be an deterministic attempt. Efficacy of every implementation inside the system is important. Feedbacks will have affects for self regulation and self evolution of this cybernetic environment. Every time it is started, audiences will observe another possibility of interactions of known events as a sonic representation.
To design a self creative model of a s